requirements.txt
openai
python-dotenv
docstring_parser